stabilize landmarks to maybe prevent infinite update loop in toc and also prevent a bunch of expensive computation on every editor keystroke for posts
